+++ tools/NTangle.Template/content/AppName.Publisher/Program.cs
@@ -1,5 +1,4 @@
-﻿using CoreEx.Configuration;
-using CoreEx.Events;
+﻿using CoreEx.Events;
 using DbEx;
 #if (implement_dbprovider_sqlserver)
 using Microsoft.Data.SqlClient;
@@ -34,10 +33,8 @@ namespace AppName.Publisher
             Host.CreateDefaultBuilder(args)
                 .ConfigureServices((hostContext, services) =>
                 {
-                    var settings = new DefaultSettings(hostContext.Configuration);
-
-                    services.AddDefaultSettings()
-                            .AddLogging(configure => configure.AddConsole())
+                   services.AddDefaultSettings()
+                            .AddLogging(b => b.AddSimpleConsole())
 #if (implement_dbprovider_sqlserver)
                             .AddDatabase(sp => new SqlServerDatabase(() => new SqlConnection(hostContext.Configuration.GetConnectionString("SqlDb"))))
 #endif
@@ -50,8 +47,8 @@ namespace AppName.Publisher
                             .AddGeneratedEventOutboxSender();
 
                     // Adds the CDC-hosted service(s) including orchestrator services, and specified EventOutbox dequeue/send service.
-                    services.AddGeneratedCdcHostedServices(settings)
-                            .AddEventOutboxHostedService(settings, sp => new EventOutboxDequeue(sp.GetRequiredService<IDatabase>(), new LoggerEventSender(sp.GetRequiredService<ILogger<LoggerEventSender>>()), sp.GetRequiredService<ILogger<EventOutboxDequeue>>()))
+                    services.AddGeneratedCdcHostedServices()
+                            .AddEventOutboxHostedService(sp => new EventOutboxDequeue(sp.GetRequiredService<IDatabase>(), new LoggerEventSender(sp.GetRequiredService<ILogger<LoggerEventSender>>()), sp.GetRequiredService<ILogger<EventOutboxDequeue>>()))
                             .AddGeneratedOrchestratorServices()
                             .AddFileLockSynchronizer();
                 });